# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: OneMore API
  version: 0.1.0
  description: OneMore fitness platform REST API v1
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /workouts/sessions/{sessionId}:
    get:
      summary: Get workout session detail with sets
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
  /workouts/sessions/{sessionId}/sets:
    put:
      summary: Upsert a set log for an exercise execution
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
  /workouts/sessions/{sessionId}/exercises:
    post:
      summary: Add an exercise to a free workout session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
  /workouts/sessions/{sessionId}/complete:
    post:
      summary: Complete a workout and advance program rotation
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
  /workouts/sessions/{sessionId}/abandon:
    post:
      summary: Abandon an in-progress workout session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
...